python_requirement.txt
selenium
pytest
pytest-xdist
pytest-html
concurrencytest
pylint
pyyaml